chore: Extract test-suite management to doctest/parts/test_suite.h
